backend/requirements.txt
fastapi
uvicorn[standard]
PyJWT
passlib==1.7.4
bcrypt==4.0.1
pydantic
python-multipart
itsdangerous
